Adding replicate variance script
This script uses the 1/n * the sum of squared differences from the mean for each set of replicates of each gene in each experiment, the ij variance over k replicates. The variance of the baseline normal, Tau^2 here, is set by (1.4826*MAD)^2 of the gene means across replicates in each experiment. The 1.4826 is the consistency scale factor for normally distributed data. The bounds of the uniform distributions used are the UMVUE's for the min and max at the gene level, i.e. Dapk has the same uniform bounds across all replicates in all experiments, and they are unique to Dapk. This script needs integrated into the TernaryNet workflow to enable likelihood scoring of sampled networks.